Делает src/components/reviews/reviews.jsx функциональным компонентом и переносит загрузку comments в movie-page.jsx
