+++ src/components/reviews/reviews.jsx
@@ -1,37 +1,26 @@
-import React, {PureComponent} from "react";
+import React from "react";
 import PropTypes from "prop-types";
 import {connect} from "react-redux";
 import ReviewsColumn from "../reviews-column/reviews-column.jsx";
 import {getComments} from "../../reducer/data/selectors.js";
-import {Operation as OperationData} from "../../reducer/data/data.js";
 
-class Reviews extends PureComponent {
-  constructor(props) {
-    super(props);
+
+const Reviews = (props) => {
+
+  const {reviews} = props;
+  if (reviews) {
+    const reviewsColumns1 = reviews.slice(0, Math.ceil(reviews.length / 2));
+    const reviewsColumns2 = reviews.slice(Math.ceil(reviews.length / 2));
+
+    return (
+      <div className="movie-card__reviews movie-card__row">
+        {reviewsColumns1.length > 0 ? <ReviewsColumn reviewsColumns={reviewsColumns1} /> : ``}
+        {reviewsColumns2.length > 0 ? <ReviewsColumn reviewsColumns={reviewsColumns2} /> : ``}
+      </div>
+    );
   }
-
-  componentDidMount() {
-    const {id, onGetComments} = this.props;
-    onGetComments(id);
-  }
-
-  render() {
-    const {reviews} = this.props;
-    if (reviews) {
-      const reviewsColumns1 = reviews.slice(0, Math.ceil(reviews.length / 2));
-      const reviewsColumns2 = reviews.slice(Math.ceil(reviews.length / 2));
-
-      return (
-        <div className="movie-card__reviews movie-card__row">
-          {reviewsColumns1.length > 0 ? <ReviewsColumn reviewsColumns={reviewsColumns1} /> : ``}
-          {reviewsColumns2.length > 0 ? <ReviewsColumn reviewsColumns={reviewsColumns2} /> : ``}
-        </div>
-      );
-    }
-    return null;
-  }
-
-}
+  return null;
+};
 
 
 Reviews.propTypes = {
@@ -49,21 +38,13 @@ Reviews.propTypes = {
       id: PropTypes.string.isRequired,
     })).isRequired,
   ]),
-  onGetComments: PropTypes.func.isRequired,
 };
-
 
 const mapStateToProps = (state) => ({
   reviews: getComments(state),
 });
 
-const mapDispatchToProps = (dispatch) => ({
-  onGetComments(id) {
-    dispatch(OperationData.loadComments(id));
-  },
-});
-
-
 export {Reviews};
 
-export default connect(mapStateToProps, mapDispatchToProps)(Reviews);
+export default connect(mapStateToProps)(Reviews);
+
